Cleanup and format code to satisfy code inspection plugins and add PR check for plugin
