+++ kt/plugins/godot-intellij-plugin/build.gradle.kts
@@ -1,7 +1,5 @@
-import io.gitlab.arturbosch.detekt.Detekt
 import org.jetbrains.changelog.closure
 import org.jetbrains.changelog.markdownToHTML
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     // Java support
