+++ kt/plugins/godot-intellij-plugin/detekt-config.yml
@@ -6,3 +6,19 @@ formatting:
     continuationIndentSize: 8
   ParameterListWrapping:
     indentSize: 8
+  MaximumLineLength:
+    maxLineLength: 240
+  CommentSpacing:
+    active: false
+
+style:
+  MaxLineLength:
+    maxLineLength: 240
+  ForbiddenComment:
+    values: [ 'STOPSHIP:' ] # explicit to allow Todo and fixme comments
+
+complexity:
+  ComplexMethod:
+    ignoreSingleWhenExpression: true
+  TooManyFunctions:
+    excludes: [ '**/godot/intellij/plugin/inspections/registration/inspector/property/PropertyHintAnnotationChecker.kt' ]
