+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/inspections/registration/inspector/function/RegisterFunctionInspection.kt
@@ -1,4 +1,4 @@
-package godot.intellij.plugin.inspections.registration.inspector
+package godot.intellij.plugin.inspections.registration.inspector.function
 
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
@@ -7,7 +7,7 @@ import godot.intellij.plugin.inspections.registration.quickfix.NotificationFunct
 import org.jetbrains.kotlin.idea.inspections.AbstractKotlinInspection
 import org.jetbrains.kotlin.idea.util.findAnnotation
 import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.psi.*
+import org.jetbrains.kotlin.psi.namedFunctionVisitor
 import org.jetbrains.kotlin.psi.psiUtil.containingClass
 
 class RegisterFunctionInspection : AbstractKotlinInspection() {
@@ -16,13 +16,17 @@ class RegisterFunctionInspection : AbstractKotlinInspection() {
     override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
         return namedFunctionVisitor { ktNamedFunction ->
             if (
-                ktNamedFunction.containingClass()?.findAnnotation(FqName("godot.annotation.RegisterClass")) != null
-                && notificationFunctions.contains(ktNamedFunction.name)
-                && ktNamedFunction.findAnnotation(FqName("godot.annotation.RegisterFunction")) == null
+                ktNamedFunction.containingClass()?.findAnnotation(FqName("godot.annotation.RegisterClass")) != null &&
+                notificationFunctions.contains(ktNamedFunction.name) &&
+                ktNamedFunction.findAnnotation(FqName("godot.annotation.RegisterFunction")) == null
             ) {
                 holder.registerProblem(
                     ktNamedFunction.navigationElement,
-                    "Overridden notification function which is not registered will not be called by Godot. Using notification functions for other purposes than to be called from Godot is considered a bad practise. Either register it or move your logic to a custom function you defined",
+                    """
+                        Overridden notification function which is not registered will not be called by Godot.
+                        Using notification functions for other purposes than to be called from Godot is considered a bad practise.
+                        Either register it or move your logic to a custom function you defined
+                    """.trimIndent(),
                     ProblemHighlightType.WARNING,
                     notificationFunctionNotRegisteredQuickFix
                 )
@@ -30,8 +34,8 @@ class RegisterFunctionInspection : AbstractKotlinInspection() {
         }
     }
 
+    // TODO: find a better way of checking all -> maybe add godot-library as a dependency and query notification functions through reflection or generate during compilation of the plugin
     private val notificationFunctions = listOf(
-        //TODO: find a better way of checking all -> maybe add godot-library as a dependency and query notification functions through reflection or generate during compilation of the plugin
         "_ready",
         "_enter_tree",
         "_exit_tree",
