+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/inspections/registration/quickfix/RegisterSignalMutabilityQuickFix.kt
@@ -20,11 +20,12 @@ class RegisterSignalMutabilityQuickFix : LocalQuickFix {
             val factory = KtPsiFactory(project)
             descriptor.psiElement.replace(factory.createValKeyword())
         } else {
+            val propertyName = descriptor.psiElement.nextSibling.text
             Notifications.Bus.notify(
                 NOTIFICATION_GROUP.value
                     .createNotification(
                         "@RegisterSignal Quick Fix",
-                        "Could not change the mutability of property ${descriptor.psiElement.nextSibling.text}. Change it manually",
+                        "Could not change the mutability of property $propertyName. Change it manually",
                         NotificationType.ERROR,
                         null
                     )
